lint

+++ pandas/tests/extension/base/ops.py
@@ -264,8 +264,9 @@ class BaseUnaryOpsTests(BaseOpsUtil):
             alt = ufunc(data)
             tm.assert_extension_array_equal(result, alt)
 
+
 class BaseRoundingTests:
     def test_round(self, data):
         result = pd.Series(data).round()
-        expected = pd.Series(np.round(data), dtype = data.dtype)
+        expected = pd.Series(np.round(data), dtype=data.dtype)
         tm.assert_series_equal(result, expected)
